services/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["holma91 <alexanderholmberg91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
svix = "^1.4.12"
alembic = "^1.11.1"
pyhumps = "^3.8.0"
sqlalchemy = "^2.0.19"
boto3 = "^1.28.25"
psycopg2-binary = "^2.9.7"
firebase-admin = "^6.2.0"
black = "^23.7.0"



[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
